refactor 24.10.10 syb 피드백 반영
