--- src/main/java/calculator/domain/ErrorMessage.java
@@ -1,9 +1,0 @@
-package calculator.domain;
-
-public class ErrorMessage {
-    public static final String BLINK_INPUT_ERROR = "입력된 값이 없습니다.";
-    public static final String INPUT_IS_NOT_NUMBER = "입력된 값이 숫자가 아닙니다.";
-    public static final String INPUT_IS_NOT_OPERATOR_SYMBOL = "사칙연산 기호가 아닙니다.";
-    public static final String DIVDE_ZERO_ERROR = "0으로 나눌수 없습니다.";
-    public static final String NOTHING_TO_CACULATE = "계산 할 수 없습니다.";
-}
