+++ src/main/java/calculator/domain/OperationHandler.java
@@ -1,28 +1,24 @@
 package calculator.domain;
 
+import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.Queue;
-import java.util.Set;
-
-import static calculator.domain.ErrorMessage.*;
 
 public class OperationHandler {
-    private static final Set<String> OPERATION_SET = Set.of("+", "-", "*", "/");
-    private Queue<String> operations;
-
+    public static final String INPUT_IS_NOT_NUMBER = "입력된 값이 숫자가 아닙니다.";
+    public static final String INPUT_IS_NOT_OPERATOR_SYMBOL = "사칙연산 기호가 아닙니다.";
+    public static final String NOTHING_TO_CACULATE = "계산 할 수 없습니다.";
+    private final Queue<String> operations;
 
     public OperationHandler(String[] operationArrays) {
-        this.operations = new LinkedList<>();
-        for (String operation : operationArrays) {
-            operations.add(operation);
-        }
+        this.operations = new LinkedList<>(Arrays.asList(operationArrays));
     }
 
     public int calculate() {
         int calc = convertStringToInt(operations.poll());
         while (!operations.isEmpty()) {
             validateCalcSize(operations);
-            Operation operation = new Operation(calc, validateOperationSymbol(operations.poll()), convertStringToInt(operations.poll()));
+            Operation operation = new Operation(calc, operations.poll(), convertStringToInt(operations.poll()));
             calc = operation.execute();
         }
         return calc;
@@ -40,12 +36,5 @@ public class OperationHandler {
         } catch (NumberFormatException ex) {
             throw new IllegalArgumentException(INPUT_IS_NOT_NUMBER);
         }
-    }
-
-    private String validateOperationSymbol(String input) {
-        if (!OPERATION_SET.contains(input)) {
-            throw new IllegalArgumentException(INPUT_IS_NOT_OPERATOR_SYMBOL);
-        }
-        return input;
     }
 }
