+++ src/test/java/calculator/CalculatorTest.java
@@ -1,11 +1,9 @@
 package calculator;
 
 import static calculator.domain.Calculator.*;
-import static calculator.domain.ErrorMessage.*;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-
+import static org.junit.jupiter.api.Assertions.assertAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -18,9 +16,11 @@ class CalculatorTest {
     void 문자열_분리() {
         String input = "4 + 6";
         String calculateArray[] = splitInput(input);
-        assertThat(calculateArray[0]).isEqualTo("4");
-        assertThat(calculateArray[1]).isEqualTo("+");
-        assertThat(calculateArray[2]).isEqualTo("6");
+        assertAll(
+                () -> assertThat(calculateArray[0]).isEqualTo("4"),
+                () -> assertThat(calculateArray[1]).isEqualTo("+"),
+                () -> assertThat(calculateArray[2]).isEqualTo("6")
+        );
     }
 
     @DisplayName("입력받은 문자가 빈 공백이거나 null일 경우 에러를 발생한다.")
@@ -32,5 +32,4 @@ class CalculatorTest {
         }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(BLINK_INPUT_ERROR);
     }
-
 }
